refactor(clawdstrike): extract async_config/broker/origins/resolver/settings from policy.rs
Pure relocation of leaf schema groups into sibling modules:
- async_config: TimeoutBehavior, AsyncExecutionMode, Async*PolicyConfig
- broker: BrokerMethod, BrokerConfig, BrokerProviderPolicy
- origins: OriginDefaultBehavior, OriginsConfig, OriginProfile, OriginMatch,
  OriginDataPolicy, OriginBudgets, BridgePolicy, BridgeTarget
- resolver: PolicyCustomGuardSpec, PolicyLocation, ResolvedPolicySource,
  PolicyResolver, LocalPolicyResolver (+ private default_true/default_json_object)
- settings: CustomGuardSpec, PolicySettings, VerificationSettings
  (+ private defaults, pub(crate) merge_verification_settings for merge.rs)

mod.rs re-exports the full public surface explicitly (no globs). The inline
test module gains an explicit use crate::origin::{...} that was previously
pulled in via the module-level import. No logic edits.
